feature：清理ani相关告警

+++ interfaces/ets/ani/hitracechain/src/hitrace_chain_ani.cpp
@@ -193,7 +193,7 @@ static void ParseHiTraceId(ani_env *env, ani_object HiTraceIdAni, HiTraceId& tra
     }
     uint32_t flags = UINT32_T_PRO_DEFAULT_VALUE;
     if (!HiTraceChainAniUtil::IsRefUndefined(env, flagsRef)) {
-        flags = HiTraceChainAniUtil::ParseNumberValueInt32(env, flagsRef);
+        flags = static_cast<uint32_t>(HiTraceChainAniUtil::ParseNumberValueInt32(env, flagsRef));
         traceId.SetFlags(flags);
     }
 }
